+++ Drop.h
@@ -10,7 +10,7 @@ public:
     bool isCaught;
     
     Drop(string fname, Type type, int x, int gameSpeed);
-    void update(long elapsed);
+    void update(Uint32 elapsed);
     bool isAlive();
 };
